// node/configure/type/sitemap-create.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type NoName2 = string;
export type Htpasswd = string;
/**
 * https://www.sitemaps.org/protocol.html
 */
export type URL = number;
export type View = string;

export interface NoName {
  auth: NoName1;
  url_limit: URL;
  view_path: View;
}
export interface NoName1 {
  realm: NoName2;
  htpasswd_file: Htpasswd;
  json_401: JSON;
}
export interface JSON {
  [k: string]: unknown;
}